smart-student-agent-assistant/pyproject.toml
[project]
name = "smart-student-agent-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
